+++ sourcing-new/src/main/scala/ch/epfl/bluebrain/nexus/sourcingnew/processor/EventSourceConfig.scala
@@ -1,4 +1,4 @@
-package ch.epfl.bluebrain.nexus.sourcingnew.eventsource
+package ch.epfl.bluebrain.nexus.sourcingnew.processor
 
 import akka.util.Timeout
